.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
pandas
matplotlib
python-pptx
pillow
python-docx
